New Automation/pyproject.toml
[tool.poetry]
name = "pharma_news"
version = "0.0.1"
description = ""
authors = ["Wais data"]
packages = [{include = "pharma_news"}]

[tool.poetry.dependencies]
python = "^3.8.5"
python-decouple = "^3.3"
muttlib = "^1.0.4"
apache-airflow = "^2.3"
mlflow = "^1.20.2"
psycopg2-binary = "^2.9.1"
MarkupSafe = "1.1.1"
gnews = "0.2.7"
newspaper3k = "0.2.8"
feedparser = "6.0.8"
pangres = "^2.2.3"
azure-storage-blob = "^12.9.0"
pymongo = {version = "3.12.2", extras = ["srv"]}
google-api-python-client = "^2.57.0"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]
pdbpp = "^0.10.2"
pylint = "^2.6.0"
pre-commit = "^2.7.1"
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.6.4"
pre-commit-hooks = "^3.3.0"
pytest = "^6.1.1"
matplotlib = "^3.3.3"
jupyter = "^1.0.0"
nbconvert = "^6.0.7"
shellcheck-py = "^0.7.1.1"
jupyterlab = "^3.1.7"
pandas-vet = "^0.2.2"

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[tool.black]
line-length = 88
skip-string-normalization = true
multi_line_output = 3


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


